Correction des tests via POSTMAN mise en place Frontend debut

+++ backend/systeme-reservation-jo/src/main/java/com/example/systeme_reservation_jo/repository/EvenementRepository.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface EvenementRepository extends JpaRepository<Evenement, Integer> {
+public interface EvenementRepository extends JpaRepository<Evenement, Long> {
 
     // Recherche par titre (méthode query method)
     List<Evenement> findByTitre(String titre);
